% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascot_trial.R
\name{ascot_trial}
\alias{ascot_trial}
\title{Run ASCOT trial simulation}
\usage{
ascot_trial(
  b,
  n_seq,
  M0 = rep(0, 5),
  S0 = diag(c(2^2, 2^2, 2^2, 0.5^2, 2^2), 5),
  delta = 0.2,
  effective_thres = 0.975,
  equivalent_thres = 0.9,
  ineffective_thres = 0.1,
  mod_effective_thres = 0.05,
  best_thres = 0.95,
  rar_control = FALSE,
  rar_best = TRUE,
  perm_drop = TRUE,
  early_t = 7,
  trans_prob = 0.75,
  mc_draw = 20000,
  return_all = FALSE,
  ...
)
}
\description{
Run ASCOT trial simulation
}
