% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascot_trial5.R
\name{ascot_trial5}
\alias{ascot_trial5}
\title{Run updated ASCOT trial simulation}
\usage{
ascot_trial5(
  b,
  n_seq,
  M0 = rep(0, 7),
  S0 = diag(c(100, rep(2.5^2, 6))),
  delta = log(1.1),
  effective_thres = 0.99,
  futility_thres = 0.95,
  inferior_thres = 0.01,
  superior_thres = 0.99,
  min_any = TRUE,
  rar_scale = 0.5,
  rar_use_ss = TRUE,
  min_rar = rep(1/5, 3),
  perm_drop = TRUE,
  use_mwud = TRUE,
  early_t = 0,
  mc_draw = 20000,
  return_all = FALSE,
  ...
)
}
\arguments{
\item{b}{The true log-odds parameter}

\item{n_seq}{The sequence of interim analysis in terms of sample size}

\item{S0}{The prior covariance matrix}

\item{delta}{The reference log-odds value for futility}

\item{effective_thres}{The threshold for declaring a treatment effective}

\item{futility_thres}{The threshold for declaring a treatment futile}

\item{inferior_thres}{The threshold for declaring inferiority, scaled by number of arms in domain}

\item{superior_thres}{The threshold for declaring superiority, scaled by number of arms in domain}

\item{rar_scale}{Scaling factor for RAR}

\item{rar_use_ss}{Use sample size in RAR}

\item{min_rar}{The minimum allocation to an arm in each domain (e.g. if Pr(alloc=1)<min_rar then set Pr(alloc=1)=min_rar)}

\item{perm_drop}{Permanently drop arms or only suspend them}

\item{use_mwud}{Use Mass-Weighted Urn for randomisation or just simple randomisation}

\item{early_t}{Timing of early event}

\item{mc_draw}{The number of Monte Carlo draws to use for computing posterior quantities}

\item{return_all}{Return all interim results}

\item{...}{Other arguments}

\item{ineffective_thres}{The threshold for declaring harm}

\item{equivalent_thres}{The threshold for declaring a treatment equivalent to SoC}

\item{scale_sup_thres}{Scale the superiorirty threshold by number of arms in domain?}

\item{rar_control}{Indicate if control should be included in RAR or fixed to 1/(num active arms)}

\item{rar_best}{Indicate if RAR should be proportional to Pr(best) or Pr(effective)}

\item{use_optimal}{Use optimal allocation ratios}
}
\description{
Run updated ASCOT trial simulation
}
