% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{colwise_mult}
\alias{colwise_mult}
\title{Multiply every column combination between two matrices}
\usage{
colwise_mult(A, B, sep = ".")
}
\arguments{
\item{A}{matrix 1}

\item{B}{matrix 2}
}
\value{
Matrix of interactions between A and B
}
\description{
Function is used for generating interactions between domain matrices.
Note that terms on matrix A change first, i.e. `out[, 1] = A[, 1]B[, 1]`, `out[, 2] = A[, 2]B[,1]` etc.
}
