% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{correct_min_rar_any}
\alias{correct_min_rar_any}
\title{Correct RAR to achieve minimum across all interventions}
\usage{
correct_min_rar_any(p, min_p, act, X, tol = 1e-04, maxiter = 20)
}
\arguments{
\item{p}{The current regimen allocations}

\item{min_p}{The minimum allocation to any active arm in the domain}

\item{act}{Active intervention status}

\item{X}{The indicator design for regimens}

\item{tol}{The tolerance level}

\item{maxiter}{The maximum iterations}
}
\description{
Correct RAR to achieve minimum across all interventions
}
