% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{correct_min_rar_soc}
\alias{correct_min_rar_soc}
\title{Correct RAR to achieve minimum on SoC}
\usage{
correct_min_rar_soc(p, min_p, act, X, tol = 1e-04, maxiter = 20)
}
\arguments{
\item{p}{The current regimen allocations}

\item{min_p}{The minimum allocation to SoC}

\item{X}{The indicator design for regimens}

\item{tol}{The tolerance level}

\item{maxiter}{The maximum iterations}
}
\description{
Correct RAR to achieve minimum on SoC
}
