% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{correct_rar}
\alias{correct_rar}
\title{Correct RAR}
\usage{
correct_rar(p, min_p, X, tol = 1e-04, maxiter = 20)
}
\arguments{
\item{p}{The current regimen allocations}

\item{min_p}{The targeted allocation for SoC}

\item{X}{The indicator design for regimens}

\item{tol}{The tolerance level}

\item{maxiter}{The maximum iterations}
}
\description{
Correct RAR
}
