% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascot_trial2.R
\name{tibble_arm_quantities}
\alias{tibble_arm_quantities}
\title{Group list of trial outcomes into a tibble}
\usage{
tibble_arm_quantities(dat, ...)
}
\arguments{
\item{dat}{The results of `lr_brar_trial` as a list}

\item{...}{Other arguments to `mclapply`}
}
\description{
Group list of trial outcomes into a tibble
}
