% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_brar_trial.R
\name{tibble_trial_data}
\alias{tibble_trial_data}
\title{Group list of trial outcomes into a tibble}
\usage{
tibble_trial_data(dat, ...)
}
\arguments{
\item{dat}{The results of `lr_brar_trial` as a list}

\item{...}{Other arguments to `mclapply`}
}
\description{
Group list of trial outcomes into a tibble
}
