% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascot_trial.R, R/ascot_trial2.R
\name{tibble_trial_quantities}
\alias{tibble_trial_quantities}
\title{Group list of trial outcomes into a tibble}
\usage{
tibble_trial_quantities(dat, ...)

tibble_trial_quantities(dat, ...)
}
\arguments{
\item{dat}{The results of `lr_brar_trial` as a list}

\item{...}{Other arguments to `mclapply`}
}
\description{
Group list of trial outcomes into a tibble

Group list of trial outcomes into a tibble
}
