% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{vb_mod}
\alias{vb_mod}
\title{Variational approximation to logistic regression with Normal prior.}
\usage{
vb_mod(
  X = diag(1, 4),
  y,
  n,
  M0 = rep(0, ncol(X)),
  S0 = diag(2^2, ncol(X)),
  ...
)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response events}

\item{n}{Response trials}

\item{M0}{Prior normal mean on regression parameter}

\item{S0}{Prior normal variance on regression parameter}

\item{...}{Additional arguments to `varapproxr::vb_logistic_n`}
}
\value{
A list giving results of ELBO optimised variational approximation.
}
\description{
Variational approximation to logistic regression with Normal prior.
}
