% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascot_trial2.R
\name{ascot_trial2}
\alias{ascot_trial2}
\title{Run ASCOT trial simulation}
\usage{
ascot_trial2(
  b,
  n_seq,
  delta = log(1.2),
  effective_thres = 0.99,
  equivalent_thres = 0.9,
  futility_thres = 0.01,
  ineffective_thres = 0.1,
  best_thres = 0.95,
  av_less_eff_thres = 0.01,
  rar_control = FALSE,
  rar_best = FALSE,
  perm_drop = TRUE,
  use_mwud = TRUE,
  use_optimal = FALSE,
  early_t = 7,
  h = 0.5,
  trans_prob = 0.75,
  mc_draw = 20000,
  return_all = FALSE,
  ...
)
}
\description{
Run ASCOT trial simulation
}
